% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagLaplaceDeconvolution.R
\name{MakeLaguerreMatrix}
\alias{MakeLaguerreMatrix}
\title{function MakeLaguerreMatrix}
\usage{
MakeLaguerreMatrix(a = 1/2, M)
}
\arguments{
\item{a, }{numeric, the scale of the  Laguerre functions basis}

\item{M, }{numeric, the maximal degree}
}
\value{
a constructor of the M-first Laguerre functions with scale a
}
\description{
return a constructor of the Laguerre function basis.
}
\examples{
\dontrun{

 library(LaplaceDeconv)
 # build the constructor
 bFctLagMat = MakeLaguerreMatrix(M=10)
 # compute the Laguerre function basis at times seq(0,15,l=100)
 bFctLagMat(seq(0,15,l=100))
 }
}
\author{
Y. Rozenholc and M. Pensky
}

